+++ src/next_month/test.js
@@ -2,21 +2,21 @@
 /* eslint-env mocha */
 
 var assert = require('power-assert')
-var nextMonth = require('./')
+var getNextMonth = require('./')
 
-describe('nextMonth', function () {
+describe('getNextMonth', function () {
   it('returns the next month of the given date', function () {
-    var result = nextMonth(new Date(2014, 11 /* Dev */, 25))
+    var result = getNextMonth(new Date(2014, 11 /* Dev */, 25))
     assert(result === 0)
   })
 
   it('accepts a string', function () {
-    var result = nextMonth(new Date(2014, 6 /* Jul */, 2).toISOString())
+    var result = getNextMonth(new Date(2014, 6 /* Jul */, 2).toISOString())
     assert(result === 7)
   })
 
   it('accepts a timestamp', function () {
-    var result = nextMonth(new Date(2014, 3 /* Apr */, 2).getTime())
+    var result = getNextMonth(new Date(2014, 3 /* Apr */, 2).getTime())
     assert(result === 4)
   })
 })